scanner/requirements.txt
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
lxml==5.3.0
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pycparser==2.22
pypdfium2==4.30.1
pytesseract==0.3.13
python-docx==1.1.2
typing_extensions==4.12.2
